+++ rootly_sdk/models/new_alert_route_data_attributes_rules_item_condition_groups_item_conditions_item_conditionable_type.py
@@ -10,8 +10,10 @@ NEW_ALERT_ROUTE_DATA_ATTRIBUTES_RULES_ITEM_CONDITION_GROUPS_ITEM_CONDITIONS_ITEM
 
 
 def check_new_alert_route_data_attributes_rules_item_condition_groups_item_conditions_item_conditionable_type(
-    value: str,
-) -> NewAlertRouteDataAttributesRulesItemConditionGroupsItemConditionsItemConditionableType:
+    value: str | None,
+) -> NewAlertRouteDataAttributesRulesItemConditionGroupsItemConditionsItemConditionableType | None:
+    if value is None:
+        return None
     if (
         value
         in NEW_ALERT_ROUTE_DATA_ATTRIBUTES_RULES_ITEM_CONDITION_GROUPS_ITEM_CONDITIONS_ITEM_CONDITIONABLE_TYPE_VALUES