+++ rootly_sdk/models/new_alert_route_data_attributes_rules_item_condition_groups_item_conditions_item_property_field_condition_type.py
@@ -26,8 +26,10 @@ NEW_ALERT_ROUTE_DATA_ATTRIBUTES_RULES_ITEM_CONDITION_GROUPS_ITEM_CONDITIONS_ITEM
 
 
 def check_new_alert_route_data_attributes_rules_item_condition_groups_item_conditions_item_property_field_condition_type(
-    value: str,
-) -> NewAlertRouteDataAttributesRulesItemConditionGroupsItemConditionsItemPropertyFieldConditionType:
+    value: str | None,
+) -> NewAlertRouteDataAttributesRulesItemConditionGroupsItemConditionsItemPropertyFieldConditionType | None:
+    if value is None:
+        return None
     if (
         value
         in NEW_ALERT_ROUTE_DATA_ATTRIBUTES_RULES_ITEM_CONDITION_GROUPS_ITEM_CONDITIONS_ITEM_PROPERTY_FIELD_CONDITION_TYPE_VALUES