+++ rootly_sdk/models/new_alerts_source_data_attributes_sourceable_attributes_type_0_field_mappings_attributes_item_field.py
@@ -24,8 +24,10 @@ NEW_ALERTS_SOURCE_DATA_ATTRIBUTES_SOURCEABLE_ATTRIBUTES_TYPE_0_FIELD_MAPPINGS_AT
 
 
 def check_new_alerts_source_data_attributes_sourceable_attributes_type_0_field_mappings_attributes_item_field(
-    value: str,
-) -> NewAlertsSourceDataAttributesSourceableAttributesType0FieldMappingsAttributesItemField:
+    value: str | None,
+) -> NewAlertsSourceDataAttributesSourceableAttributesType0FieldMappingsAttributesItemField | None:
+    if value is None:
+        return None
     if (
         value
         in NEW_ALERTS_SOURCE_DATA_ATTRIBUTES_SOURCEABLE_ATTRIBUTES_TYPE_0_FIELD_MAPPINGS_ATTRIBUTES_ITEM_FIELD_VALUES