+++ rootly_sdk/models/update_alerts_source_data_attributes_resolution_rule_attributes_type_0_conditions_attributes_item_kind.py
@@ -13,8 +13,10 @@ UPDATE_ALERTS_SOURCE_DATA_ATTRIBUTES_RESOLUTION_RULE_ATTRIBUTES_TYPE_0_CONDITION
 
 
 def check_update_alerts_source_data_attributes_resolution_rule_attributes_type_0_conditions_attributes_item_kind(
-    value: str,
-) -> UpdateAlertsSourceDataAttributesResolutionRuleAttributesType0ConditionsAttributesItemKind:
+    value: str | None,
+) -> UpdateAlertsSourceDataAttributesResolutionRuleAttributesType0ConditionsAttributesItemKind | None:
+    if value is None:
+        return None
     if (
         value
         in UPDATE_ALERTS_SOURCE_DATA_ATTRIBUTES_RESOLUTION_RULE_ATTRIBUTES_TYPE_0_CONDITIONS_ATTRIBUTES_ITEM_KIND_VALUES