+++ rootly_sdk/models/update_communications_group_data_attributes_communication_group_conditions_type_0_item_property_type.py
@@ -16,8 +16,10 @@ UPDATE_COMMUNICATIONS_GROUP_DATA_ATTRIBUTES_COMMUNICATION_GROUP_CONDITIONS_TYPE_
 
 
 def check_update_communications_group_data_attributes_communication_group_conditions_type_0_item_property_type(
-    value: str,
-) -> UpdateCommunicationsGroupDataAttributesCommunicationGroupConditionsType0ItemPropertyType:
+    value: str | None,
+) -> UpdateCommunicationsGroupDataAttributesCommunicationGroupConditionsType0ItemPropertyType | None:
+    if value is None:
+        return None
     if (
         value
         in UPDATE_COMMUNICATIONS_GROUP_DATA_ATTRIBUTES_COMMUNICATION_GROUP_CONDITIONS_TYPE_0_ITEM_PROPERTY_TYPE_VALUES